Add CircleCI job using REST service backed by CRUD-only K8s
This change adds a CircleCI workflow that uses the test-helper package
that has been published to nuodb-cp-releases, to allow the tests to run
much faster and enable us to increase test coverage.

The test-helper package includes the NuoDB Control Plane REST service
and code that allows a CRUD-only Kubernetes cluster to be created along
with controllers that simulate reconciliation by the NuoDB Operator.

The test themselves take only about 10 seconds to run in this
configuration, compared to about five minutes for the existing
end-to-end configuration.