Remove organization from provider config
This change removes organization (NUODB_CP_ORGANIZATION) from the
provider config. The only reason why the provider config required the
organization is to create the user name, which is simply
`<organization>/<user>`.

This served no purpose as a separate field and resulted in
inconsistencies between the provider and the nuodb-cp tool in how
NUODB_CP_USER is interpreted. nuodb-cp treats NUODB_CP_USER as the
complete user name (`<organization>/<user>`), so the same set of
environment variables would not work with both nuodb-cp and Terraform.

This change also removes the authentication check in the provider, which
does not serve any purpose since the same authentication credentials are
sent with every request. If the authentication credentials are invalid,
then REST request to inspect or manipulate the project or database
resource will fail.

This change also uses the Docker executor for the integration tests to
reduce resource usage.