Use code generation for Terraform schemas and model structs (#16)
This change removes Terraform boilerplate and duplication of information
that is already specified in the OpenAPI spec for the Control Plane REST
API and relies on code generation of the model structs and Terraform
schemas.

The openapi-generator is replaced with oapi-codegen, which has several
advantages:

- It allows extra tags to be added to generated fields on the Golang
  model structs. This allows the `tfsdk:"..."` tags to be added along
  with the `json:"..."` tags needed for interactions with the DBaaS REST
  API, so that the same model struct can be used for both the REST API
  and Terraform.
- It has a `spec` generator which embeds the OpenAPI spec and generates
  boilerplate helper code that allows the spec to be easily accessed at
  runtime. This allows us to dynamically convert from the JSONSchema
  specifications of REST models in the OpenAPI spec to Terraform
  schemas.
- It produces much less code than openapi-generator, and allows
  filtering of tags on the OpenAPI spec for the resource paths that are
  relevant (e.g. project and database). This removes irrelevant code
  having to do with service tiers, database quotas, etc.
- It can be installed using `go install` which allows easy integration
  with `go generate`.

To enable conversion from OpenAPI definitions to Terraform schemas,
several custom extensions are being added to the spec generated by the
DBaaS REST API:

- `x-tf-name` is the name of the Terraform attribute associated with a
  property.
- `x-tf-sensitve` indicates that the field is sensitive and should not be
  displayed in output.
- `x-tf-identifier` indicates that the field is an identifier and can be
  used in the configuration of a data source to identify the object to
  return.

The custom extensions are not available in any released version of the
DBaaS REST API, so this change commits an in-development version of
openapi.yaml to this repository. Once 2.4.0 of the DBaaS Control Plane
is released, we can generate from a published spec.

This change also includes refactoring changes that standardize how
Terraform resources and data sources exposing access DBaaS are defined.
Interaction with Terraform APIs is limited to GenericResource and
GenericDataSource, which can be used to implement any resource or data
source that accesses the DBaaS API.

Finally, there are a few changes in behavior introduced by this change.
Previously we were not storing the state immediately after creation or
update of a resource, but after waiting for the resource to become
ready. This was problematic because a new resource would not be
manageable via Terraform if the readiness check timed out. Now we store
the resource immediately after creation or update.