Update state after resource is ready
This change performs an update of the Terraform state after the
readiness check for create or update completes, so that the latest state
retrieved from the server is available. Previously we were only updating
the state before the polling loop where we await readiness, which is
necessary so that we can still manage the resource if the polling loop
times out, but this would leave stale state in Terraform.