Correctly handle 404 in data sources
This fixes a bug in the generic_data_source Read() method, which was copied from
the generic_resource Read() method and was removing the resource from
the state if it was not found and generating the following output when
running `terraform apply`:

```
Error: Provider produced null object

Provider "provider[\"registry.terraform.io/nuodb/nuodbaas\"]" produced a null value for data.nuodbaas_database.database_details.

This is a bug in the provider, which should be reported in the provider's own issue tracker.
```

It makes no sense to manipulate the state if the object is not found
because a data source is not managing the state of the resource.