Gate RequiresReplace plan modifier on environment variable
This change gates the behavior of the RequiresReplace plan modifier on
the environment variable NUODB_CP_ALLOW_DESTRUCTIVE_REPLACE=true.
Previously, a change to an immutable attribute would cause Terraform to
automatically delete and re-create the resource, which would cause all
database data to be lost.

With this change, the following warning is displayed before allowing the
immutable attribute change to be sent to the REST service, which will
reject it (changes to dba_password are a special case, because the DBaaS
provider suppresses dbaPassword from the request payload on updates):

```
 Warning: Immutable Attribute Change

   with nuodbaas_database.db,
   on main.tf line 13, in resource "nuodbaas_database" "db":
   13: resource "nuodbaas_database" "db" {

 A change has been made to the immutable attribute `dba_password`, which may
 be rejected by the server or ignored. In order for an immutable attribute
 change to take effect, it is necessary to delete and re-create the
 resource, which may result in data loss.

 If it is acceptable to destroy all data associated with the resource, then
 deletion and re-creation can be enabled by setting the environment variable
 `NUODB_CP_ALLOW_DESTRUCTIVE_REPLACE=true` when running `terraform apply`.

 For resources that are marked as dependencies for other resources, it may
 necessary to destroy and re-create all resources by running `terraform
 destroy` followed by `terraform apply`.

 If it is not acceptable to destroy all data associated with the resource,
 the configured value of the attribute can be reverted to the value in the
 Terraform state (`<redacted>`).
```