Allow development versions of OpenAPI spec to be used
This change avoids downloading the released version of the OpenAPI spec
if the commit is not tagged with a version, allowing us to use
development versions of the spec. If commit is tagged, then any
differences between the checked in spec and the released spec should
cause the goreleaser GitHub Action to fail, because it has a `before`
hook that invokes `make generate` followed by `make check-no-changes`.

This change also addresses linter errors having to do with conversion of
the values from the typed modifiers from the Terraform library.