Allow DBA password to be updated via Terraform
This change allows the DBA password to be updated declaratively via
Terraform. If a change is detected to the `dba_password` attribute, then
we first attempt to update the password using the `POST /dbaPassword`
sub-resource. If that fails for any reason, then we throw an error. We
specifically detect the case where the failure is due to the
sub-resource not being supported and display a message telling the user
to revert the value of the password.

This is a change in behavior, since previously we would ignore changes
to the DBA password and display a warning, unless the environment
variable `NUODB_CP_ALLOW_DESTRUCTIVE_REPLACE=true` was set, which caused
the database to be deleted and recreated. The previous behavior does not
seem desirable, and a support ticket was already filed about it by a
customer, so it seems okay to change it.

This change also includes:

- Updates the OpenAPI spec to target the latest development version of
  the Control Plane REST server. This will have to be set to the
  published version before releasing 1.1.0.
- Includes some refactoring to the way the schema is built to allow the
  specific resources that use GenericResource to supply the Terraform
  schema attributes, rather than allowing the GenericResource to convert
  the OpenAPI schema to the Terraform schema attributes. This is useful
  in case we want to include special handling of certain attributes that
  is different from the behavior associated with the various OpenAPI
  extensions (e.g. x-immutable).
- The GenericResource is updated to supply both the plan and the state,
  to allow implementer of the ResourceState interface to perform special
  update logic for certain attribute changes.