Create deployment framework for Terraform tests (#47)
This change adds a framework for deploying NuoDB Control Plane instances
in various configurations, and consolidates the CircleCI jobs that were
previously known as integration_test and e2e_test.

The integration_test job was relying on the test-helper.tgz package that
was published into the nuodb/nuodb-cp-releases repo. This has been
replaced by KWOK, which is a tool for creating K8s clusters for test
purposes. Unlike EnvTest, KWOK includes the controller-manager, which
enables garbage collection and scheduling of workloads like Deployments
and StatefulSets. This allows the simulated reconciliation logic that
was previously being used to be replaced by real reconciliation logic
being performed by an instance of the NuoDB Operator.

This change also decouples various aspects of the test environment that
were previously conflated together based on the E2E_TEST environment
variable.  CONTAINER_SCHEDULING_ENABLED is used to skip tests that would
take too long in a true end-to-end configuration, or inject
configuration to speed up the test.  WEBHOOKS_ENABLED is used to skip
tests that exercise invalid configurations, which may be rejected by the
server in a way that prevents us from achieving the desired Terraform
test coverage.