Add testing on version 2.5.0
This change adds CircleCI jobs for 2.5.0. The 2.4.1 jobs will be
maintained as a "lowest supported version" test.

The upgrade to 2.5.0 exposed an issue with state reporting where a
project or database is prematurely reported as Available and then
transitions to Modifying. This change adds a workaround as we wait for a
fix in the Control Plane.