Allow OpenAPI description overrides and fix dba_password description
The OpenAPI spec correctly says that the `dbaPassword` field can only
supplied when creating a database using `PUT /databases/org/proj/db`,
but since Terraform automatically invokes `POST
/databases/org/proj/db/dbaPassword` when it detects changes to
`dba_password`, this description is not relevant to the `dba_password`
attribute.

This change allows overrides to be defined on OpenAPI schema properties
and specifies an override on the `dbaPassword` property, which is used
for the `dba_password` Terraform attribute.