Run tests against latest dev version (main) using KWOK
This change allows us to test against the latest development version of
the DBaaS Control Plane using KWOK. This is useful because KWOK gives us
the highest test coverage of all test drivers and allows us to test
features as soon as they have been pushed to main on
nuodb-control-plane, even if they have not been deployed to the test
cluster used by the "External DBaaS" variant.

This change also restores some test coverage that was lost by only
running `terraform validate` on the example configs. If the REST server
is the latest dev version, then we can enable full testing of the
example configs, since all resources and data sources should be
supported.