Add overrides to database schema to require full backup name (#71)
Terraform fails when it detects changes to configuration values, even if
the change is to an equivalent form. This occurs when a partial backup
name is supplied, which is normalized by the server to resolve missing
path components from the database.

To avoid Terraform errors, require the `restore_from.backup` attribute
to specify the fully-qualified name.